examples/tpcdi/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[tool.setuptools.package-data]
"*" = []

[project]
name = "interlace-tpcdi-benchmark"
version = "0.1.0"
description = "TPC-DI Benchmark for Interlace - Data Integration"
requires-python = ">=3.12"
dependencies = [
    "interlace",
    "ibis-framework[duckdb]>=9.0.0",
    "pandas>=2.0.0",
]

[tool.uv.sources]
interlace = { workspace = true }


